requirements.txt
PyMuPDF==1.22.5
pdfplumber
pandas
streamlit
google-generativeai
openai
python-dotenv